requirements.txt
absl-py==0.10.0
astor==0.8.1
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.1
gast==0.3.3
gevent==20.9.0
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.32.0
gunicorn==19.9.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.17.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
numpy==1.19.2
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.3
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.25.10
Werkzeug==0.15.5
wincertstore==0.2
wrapt==1.12.1
zipp==3.3.0
zope.event==4.5.0
zope.interface==5.1.2
